Don't set Content-Type for the whole view.
Liquid templates aren't just for HTML, and even if they were it's not a
renderer's place to set the Content-Type for the controller's response.

This resulted in incorrect Content-Type headers for email and for Liquid
partials inside non-HTML responses.

Tests still pass and text/html is still returned in the common case
anyway due to Rails standard behaviour.

Fixes #26.